Moved extensions to test
